pyproject.toml
[project]
name = "ai-investment-analyst"
version = "0.1.0"
description = "AI Investment Analyst data platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
  "yfinance>=0.2.54",
  "pandas>=2.2",
  "requests>=2.32",
  "finlab>=1.5.3",
  "opencc-python-reimplemented>=0.1.7",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "jinja2>=3.1",
  "python-multipart>=0.0.9"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
